Update relative link checking for `<Nav />` component
This solution doesn't feel 100% satisfactory but `Astro.site.pathname` is no longer available/respecting the value of `config.base` (I'm inclined to think that's an oversight or a bug, tbh).  I don't expect this site ever to be served from a sub-folder URL (so it could just be hardcoded/ignored), but this should be robust should the need ever arise.