Use luxon for date parsing
This addresses a complicated issue concerning timezones; essentially the yaml parsing plugin would render dates into UTC according to the timezone they are entered with (e.g. PST) and then the JS engine would format them back out according to the timezone used at runtime (which might be PST) and they could end up wrong.  This problem can be very hard to reason about, and is further exacerbated when the build is done in CD (e.g. on a GitHub Actions runner) and a third timezone is potentially in play.

Solution: specify all times in local time and read and write out time formats without timezones.